Rename KothRegisterEvent to KothHookEvent Change package structure
