+++ src/main/java/fr/maxlego08/koth/api/FactionListener.java
@@ -1,4 +1,4 @@
-package fr.maxlego08.koth;
+package fr.maxlego08.koth.api;
 
 import java.util.List;
