+++ src/main/java/fr/maxlego08/koth/api/event/events/KothHookEvent.java
@@ -1,0 +1,9 @@
+package fr.maxlego08.koth.api.event.events;
+
+public class KothHookEvent {
+
+	public void setFactionListener(Object obj){
+		
+	}
+
+}
