--- src/main/java/fr/maxlego08/koth/event/KothRegisterEvent.java
@@ -1,9 +1,0 @@
-package fr.maxlego08.koth.event;
-
-public class KothRegisterEvent {
-
-	public void setFactionListener(Object obj){
-		
-	}
-
-}
